+++ java-sample/src/test/resources/io/github/pixee/maven/operator/test/broken-pom.xml
